# api_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "The Third Voice AI",
        "description": "AI-powered communication assistant for difficult conversations - Healing families through better communication",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/messages/quick-transform": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Quick Transform",
                "description": "Quick message transformation without authentication\n\nTransforms messages to be more constructive and healing",
                "operationId": "quick_transform_api_messages_quick_transform_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuickMessageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Quick Transform Api Messages Quick Transform Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/messages/quick-interpret": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Quick Interpret",
                "description": "Quick message interpretation without authentication\n\nHelps understand what someone really means and suggests responses",
                "operationId": "quick_interpret_api_messages_quick_interpret_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuickMessageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Quick Interpret Api Messages Quick Interpret Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "QuickMessageRequest": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "contact_context": {
                        "type": "string",
                        "title": "Contact Context",
                        "default": "friend"
                    },
                    "contact_name": {
                        "type": "string",
                        "title": "Contact Name",
                        "default": "Friend"
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "QuickMessageRequest"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}